requirements.txt
# Django and web framework
Django>=4.2
django-htmx>=1.17
django-q2>=1.6

# Async HTTP client
aiohttp>=3.9

# Database (using SQLite by default, add these for production)
# psycopg2-binary>=2.9  # PostgreSQL
# mysqlclient>=2.2      # MySQL

# HTTP client (for legacy sync methods)
requests>=2.31

# System info
psutil>=5.9

# Production WSGI server
gunicorn>=21.2
whitenoise>=6.5

# PDF processing
PyMuPDF>=1.24
Pillow>=10.0

# Vector database
chromadb>=0.4

# Development
python-dotenv>=1.0
